предфинальная рабочая стадия
