+++ app.py
@@ -3,6 +3,7 @@ import logging
 from flask import Flask
 from main.main_views import main_blueprint, search_blueprint, tag_blueprint
 from bookmark.bookmark_views import bookmarks_blueprint
+from bookmark.change_bookmarks import change_bookmarks
 from api.api_views import api_blueprint
 
 app = Flask(__name__)
@@ -11,6 +12,7 @@ app.register_blueprint(main_blueprint)
 app.register_blueprint(tag_blueprint)
 app.register_blueprint(search_blueprint)
 app.register_blueprint(bookmarks_blueprint)
+app.register_blueprint(change_bookmarks)
 app.register_blueprint(api_blueprint)
 
 logging.basicConfig(filename="logs.log", encoding="utf-8", level=logging.INFO,
